+++ src/bin/cli.ts
@@ -17,11 +17,11 @@ async function main(): Promise<void> {
         .version(version);
 
     program
+        /* eslint-disable max-len */
         .command('build')
         .argument('<old_filter>', 'the relative path to the old filter')
         .argument('<new_filter>', 'the relative path to the new filter')
-        /* eslint-disable max-len */
-        .argument('<path_to_patches>', 'the relative path to the directory where the patch should be saved. The patch filename will be `<path_to_patches>/$PATCH_VERSION.patch`, where `$PATCH_VERSION` is the value of `Version` from `<old_filter>`')
+        .argument('<path_to_patches>', 'the relative path to the patches. The patch filename will be `<path_to_patches>/$PATCH`, where `$PATCH` is the "<patchName>[-<resolution>]-<epochTimestamp>-<expirationPeriod>.patch"')
         .requiredOption('-n, --name <name>', 'name of the patch file, an arbitrary string to identify the patch. Must be a string of length 1-64 with no spaces or other special characters.')
         .requiredOption('-t, --time <expirationPeriod>', 'expiration time for the diff update (the unit depends on `resolution`).')
         .option('-r, --resolution <timestampResolution>', 'is an optional flag, that specifies the resolution for both `expirationPeriod` and `epochTimestamp` (timestamp when the patch was generated). It can be either `h` (hours), `m` (minutes) or `s` (seconds). If `resolution` is not specified, it is assumed to be `h`.')