+++ tests/diff-builder.test.ts
@@ -1,12 +1,10 @@
 import { calculateChecksumSHA1 } from '../src/common/calculate-checksum';
 import { createDiffDirective } from '../src/common/diff-directive';
-import { parseTag } from '../src/common/parse-tag';
 import { TypesOfChanges } from '../src/common/types-of-change';
 import {
     detectTypeOfChanges,
     createPatch,
-    checkIfPatchIsEmpty,
-    updateDiffPathInNewFilter,
+    updateTags,
 } from '../src/diff-builder/build';
 import {
     FILE_1,
@@ -27,8 +25,7 @@ import { FILTER_2_V_1_0_0, FILTER_2_V_1_0_1, PATCH_2_1_0_0 } from './stubs/valid
 import { FILTER_3_V_1_0_0, FILTER_3_V_1_0_1, PATCH_3_1_0_0 } from './stubs/name';
 import { FILTER_WITHOUT_EMPTY_LINE, FILTER_WITH_EMPTY_LINE, FILTER_WITH_SEVERAL_EMPTY_LINES } from './stubs/new-lines';
 import { splitByLines } from '../src/common/split-by-lines';
-import { findAndUpdateTag } from '../src/diff-builder/tags';
-import { EMPTY_PATCH_1, EMPTY_PATCH_2, NOT_EMPTY_PATCH_1 } from './stubs/empty-patches';
+import { parseTag } from '../src/diff-builder/tags';
 import {
     FILTER_CHECKSUM_1_V_1_0_0,
     FILTER_CHECKSUM_1_V_1_0_1,
@@ -85,94 +82,28 @@ describe('check diff-builder', () => {
         expect(splitted[splitted.length - 1].endsWith('\n')).toBeTruthy();
     });
 
-    it('check findAndUpdateTag', () => {
-        const content = FILTER_1_V_1_0_0;
+    describe('check updateTags', () => {
+        const cases = [
+            /* eslint-disable max-len */
+            ['checks case when Diff-Path placed NOT on the first line after Checksum tag', FILTER_CHECKSUM_1_V_1_0_0, FILTER_CHECKSUM_1_V_1_0_1],
+            ['checks case when Diff-Path placed on the first line after Checksum tag', FILTER_CHECKSUM_2_V_1_0_0, FILTER_CHECKSUM_2_V_1_0_1],
+            ['checks case when there are two checksums in file', FILTER_CHECKSUM_3_V_1_0_0, FILTER_CHECKSUM_3_V_1_0_1],
+            /* eslint-enable max-len */
+        ];
 
-        const filterWithUpdatedVersion = findAndUpdateTag(
-            'Version',
-            'v9.9.9',
-            splitByLines(content),
-        );
-        expect(filterWithUpdatedVersion.join('')).toEqual(
-            content.replace(
-                '! Version: v1.0.0',
-                '! Version: v9.9.9',
-            ),
-        );
-
-        const filterWithUpdatedDiffPath = findAndUpdateTag(
-            'Diff-Path',
-            '../patches/1/1-m-28378192-60.patch',
-            splitByLines(content),
-        );
-        expect(filterWithUpdatedDiffPath.join('')).toEqual(
-            content.replace(
-                '! Diff-Path: ../patches/1/1-m-28378132-60.patch',
-                '! Diff-Path: ../patches/1/1-m-28378192-60.patch',
-            ),
-        );
-
-        // Check that original filter didn't changed.
-        expect(content).toEqual(FILTER_1_V_1_0_0);
-    });
-
-    it('check checkIfPatchIsEmpty', () => {
-        expect(checkIfPatchIsEmpty(EMPTY_PATCH_1)).toBeTruthy();
-        expect(checkIfPatchIsEmpty(EMPTY_PATCH_2)).toBeTruthy();
-        expect(checkIfPatchIsEmpty(NOT_EMPTY_PATCH_1)).toBeFalsy();
-        expect(checkIfPatchIsEmpty(NOT_EMPTY_PATCH_1)).toBeFalsy();
-    });
-
-    describe('check updateDiffPathInNewFilter', () => {
-        it('checks case when Diff-Path placed on the first line after Checksum tag', () => {
-            const filter1 = FILTER_CHECKSUM_2_V_1_0_0;
-            const filter2 = FILTER_CHECKSUM_2_V_1_0_1;
-
+        it.each(cases)('%s', (_, filter1, filter2) => {
             // Emulate changes
             const updatedFilter1 = filter1
+                .replace('! Diff-Path: ../included_filter_patch.patch', '! Diff-Path: ../included_filter_patch_2.patch')
                 .replace('Version: v1.0.0', 'Version: v1.0.1')
                 .replace('example.org', 'example.com');
 
-            const updatedFilter1WithTags = updateDiffPathInNewFilter(
-                splitByLines(updatedFilter1),
+            const updatedFilter1WithTags = updateTags(
+                updatedFilter1,
                 '../patches/1/1-m-28378192-60.patch',
             );
 
-            expect(updatedFilter1WithTags.join('')).toEqual(filter2);
-        });
-
-        it('checks case when Diff-Path placed NOT on the first line after Checksum tag', () => {
-            const filter1 = FILTER_CHECKSUM_1_V_1_0_0;
-            const filter2 = FILTER_CHECKSUM_1_V_1_0_1;
-
-            // Emulate changes
-            const updatedFilter1 = filter1
-                .replace('Version: v1.0.0', 'Version: v1.0.1')
-                .replace('example.org', 'example.com');
-
-            const updatedFilter1WithTags = updateDiffPathInNewFilter(
-                splitByLines(updatedFilter1),
-                '../patches/1/1-m-28378192-60.patch',
-            );
-
-            expect(updatedFilter1WithTags.join('')).toEqual(filter2);
-        });
-
-        it('checks case when there are two checksums in file', () => {
-            const filter1 = FILTER_CHECKSUM_3_V_1_0_0;
-            const filter2 = FILTER_CHECKSUM_3_V_1_0_1;
-
-            // Emulate changes
-            const updatedFilter1 = filter1
-                .replace('v1.0.0', 'v1.0.1')
-                .replace('another_patch.patch', 'another_patch_2.patch');
-
-            const updatedFilter1WithTags = updateDiffPathInNewFilter(
-                splitByLines(updatedFilter1),
-                '../patches/1/1-m-28378192-60.patch',
-            );
-
-            expect(updatedFilter1WithTags.join('')).toEqual(filter2);
+            expect(updatedFilter1WithTags).toEqual(filter2);
         });
     });
 
@@ -203,7 +134,9 @@ describe('check diff-builder', () => {
 
             let patch = createPatch(filter1, filter2);
 
-            const diffDirective = createDiffDirective(splitByLines(filter1), filter2, patch);
+            const oldFileDiffPath = parseTag('Diff-Path', splitByLines(filter1));
+
+            const diffDirective = createDiffDirective(oldFileDiffPath, filter2, patch);
             patch = diffDirective.concat('\n', patch);
 
             const directive = 'diff checksum:17fbb4b268e4d176fbd75fd627ba2e84b945e077 lines:4';
@@ -218,7 +151,9 @@ describe('check diff-builder', () => {
             const patch = createPatch(filter1, filter2);
             expect(patch).toEqual(PATCH_CHECKSUM_1_1_0_0);
 
-            const diffDirective = createDiffDirective(splitByLines(filter1), filter2, patch);
+            const oldFileDiffPath = parseTag('Diff-Path', splitByLines(filter1));
+
+            const diffDirective = createDiffDirective(oldFileDiffPath, filter2, patch);
             expect(diffDirective).toEqual(FILTER_CHECKSUM_1_V_1_0_1_DIFF_DIRECTIVE);
         });
     });