+++ tests/stubs/filters-with-checksum.ts
@@ -4,23 +4,23 @@ export const FILTER_CHECKSUM_1_V_1_0_0 = `! Checksum: G3J40YYEG72sOhSMJlicgw
 ! Diff-Path: ../patches/1/1-m-28378132-60.patch
 ||example.org^`;
 
-export const FILTER_CHECKSUM_1_V_1_0_1 = `! Checksum: iP3cfJ4iR+MPqRJUsy8cWw
+export const FILTER_CHECKSUM_1_V_1_0_1 = `! Checksum: bbIycvMnuGxvqqneHnvTbQ
+! Diff-Path: ../patches/1/1-m-28378192-60.patch
 ! Title: Diff Updates Simple Example List
 ! Version: v1.0.1
-! Diff-Path: ../patches/1/1-m-28378192-60.patch
 ||example.com^`;
 
 export const PATCH_CHECKSUM_1_1_0_0 = `d1 1
-a1 1
-! Checksum: iP3cfJ4iR+MPqRJUsy8cWw
-d3 3
-a5 3
-! Version: v1.0.1
+a1 2
+! Checksum: bbIycvMnuGxvqqneHnvTbQ
 ! Diff-Path: ../patches/1/1-m-28378192-60.patch
+d3 3
+a5 2
+! Version: v1.0.1
 ||example.com^`;
 
 // eslint-disable-next-line max-len
-export const FILTER_CHECKSUM_1_V_1_0_1_DIFF_DIRECTIVE = 'diff checksum:71e00a721bdb9b1da0ddb68ba23ef836cc8f8c34 lines:7';
+export const FILTER_CHECKSUM_1_V_1_0_1_DIFF_DIRECTIVE = 'diff checksum:816f1e9f77d291487c781bd559cdb36a26849618 lines:7';
 
 export const FILTER_CHECKSUM_2_V_1_0_0 = `! Checksum: ABC
 ! Diff-Path: ../patches/1/1-m-28378132-60.patch
@@ -39,13 +39,13 @@ export const FILTER_CHECKSUM_3_V_1_0_0 = `! Checksum: u/T910qA7w7mONgsMzuCOg
 ! Title: Diff Updates Simple Example List
 ! Version: v1.0.0
 ! Checksum: CCC
-! Diff-Path: ../another_patch.patch
+! Diff-Path: ../included_filter_patch.patch
 ||example.org^`;
 
-export const FILTER_CHECKSUM_3_V_1_0_1 = `! Checksum: BlDDiB3PAOSr0uK+yG78pw
+export const FILTER_CHECKSUM_3_V_1_0_1 = `! Checksum: ofZAa0I54y1eSvFOPN2k0Q
 ! Diff-Path: ../patches/1/1-m-28378192-60.patch
 ! Title: Diff Updates Simple Example List
 ! Version: v1.0.1
 ! Checksum: CCC
-! Diff-Path: ../another_patch_2.patch
-||example.org^`;
+! Diff-Path: ../included_filter_patch_2.patch
+||example.com^`;